+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [1.2.0] - 2023-08-17
 ### Added
 - Add new THINGER\_STREAMS\_READY state listener
+- Install script and service file
 
 ## [1.1.0] - 2023-06-28
 ### Added